requirements.txt
telethon
PyQt5
asyncio
emoji
python-dotenv